backend/requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.10
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
google-generativeai==0.5.4
ipython==8.26.0
